* 重新生成代码

+++ pnt-api/src/main/java/rebue/pnt/mo/PntAccountMo.java
@@ -31,7 +31,7 @@ public class PntAccountMo implements Serializable {
      *
      *    @mbg.generated 自动生成，如需修改，请删除本行
      */
-    private BigDecimal points;
+    private Long points;
 
     /**
      *    当前收益
@@ -103,7 +103,7 @@ public class PntAccountMo implements Serializable {
      *
      *    @mbg.generated 自动生成，如需修改，请删除本行
      */
-    public BigDecimal getPoints() {
+    public Long getPoints() {
         return points;
     }
 
@@ -114,7 +114,7 @@ public class PntAccountMo implements Serializable {
      *
      *    @mbg.generated 自动生成，如需修改，请删除本行
      */
-    public void setPoints(BigDecimal points) {
+    public void setPoints(Long points) {
         this.points = points;
     }
