+++ pnt-api/src/main/java/rebue/pnt/mo/PntPointsLogMo.java
@@ -3,7 +3,6 @@ package rebue.pnt.mo;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 /**
  * 积分日志
@@ -49,7 +48,7 @@ public class PntPointsLogMo implements Serializable {
      *
      *    @mbg.generated 自动生成，如需修改，请删除本行
      */
-    private BigDecimal pointsBeforeChanged;
+    private Long pointsBeforeChanged;
 
     /**
      *    改变的积分
@@ -58,7 +57,7 @@ public class PntPointsLogMo implements Serializable {
      *
      *    @mbg.generated 自动生成，如需修改，请删除本行
      */
-    private BigDecimal changedPoints;
+    private Long changedPoints;
 
     /**
      *    改变后的积分(改变后的积分=改变前的积分+改变的积分)
@@ -67,7 +66,7 @@ public class PntPointsLogMo implements Serializable {
      *
      *    @mbg.generated 自动生成，如需修改，请删除本行
      */
-    private BigDecimal pointsAfterChanged;
+    private Long pointsAfterChanged;
 
     /**
      *    改变积分的标题
@@ -201,7 +200,7 @@ public class PntPointsLogMo implements Serializable {
      *
      *    @mbg.generated 自动生成，如需修改，请删除本行
      */
-    public BigDecimal getPointsBeforeChanged() {
+    public Long getPointsBeforeChanged() {
         return pointsBeforeChanged;
     }
 
@@ -212,7 +211,7 @@ public class PntPointsLogMo implements Serializable {
      *
      *    @mbg.generated 自动生成，如需修改，请删除本行
      */
-    public void setPointsBeforeChanged(BigDecimal pointsBeforeChanged) {
+    public void setPointsBeforeChanged(Long pointsBeforeChanged) {
         this.pointsBeforeChanged = pointsBeforeChanged;
     }
 
@@ -223,7 +222,7 @@ public class PntPointsLogMo implements Serializable {
      *
      *    @mbg.generated 自动生成，如需修改，请删除本行
      */
-    public BigDecimal getChangedPoints() {
+    public Long getChangedPoints() {
         return changedPoints;
     }
 
@@ -234,7 +233,7 @@ public class PntPointsLogMo implements Serializable {
      *
      *    @mbg.generated 自动生成，如需修改，请删除本行
      */
-    public void setChangedPoints(BigDecimal changedPoints) {
+    public void setChangedPoints(Long changedPoints) {
         this.changedPoints = changedPoints;
     }
 
@@ -245,7 +244,7 @@ public class PntPointsLogMo implements Serializable {
      *
      *    @mbg.generated 自动生成，如需修改，请删除本行
      */
-    public BigDecimal getPointsAfterChanged() {
+    public Long getPointsAfterChanged() {
         return pointsAfterChanged;
     }
 
@@ -256,7 +255,7 @@ public class PntPointsLogMo implements Serializable {
      *
      *    @mbg.generated 自动生成，如需修改，请删除本行
      */
-    public void setPointsAfterChanged(BigDecimal pointsAfterChanged) {
+    public void setPointsAfterChanged(Long pointsAfterChanged) {
         this.pointsAfterChanged = pointsAfterChanged;
     }
